requirements.txt
# Runtime requirements

flask==1.1.2                    # BSD 3-Clause
Flask-SQLAlchemy==2.5.1         # BSD 3-Clause
Flask-Migrate==2.7.0            # MIT
Flask-Marshmallow==0.14.0       # MIT
marshmallow-sqlalchemy==0.25.0  # MIT
Flask-OIDC==1.4.0               # MIT

requests_oauthlib==1.3.0        # ISC

cli-ui==0.12                    # BSD 3-Clause
halo==0.0.31                    # MIT
